refactor: move helpers out of iframe/main
